[add]: Index hero initial layout
