+++ components/Homepage/Hero.tsx
@@ -1,126 +1,32 @@
 "use client"
+import { Box, Button, Container, Heading, List, ListIcon, ListItem } from '@chakra-ui/react'
+import React from 'react'
+import { HiMiniArrowRightCircle } from 'react-icons/hi2'
 
-import { useEffect, useMemo, useState } from "react";
-import Particles, { initParticlesEngine } from "@tsparticles/react";
-import {
-    type Container,
-    type ISourceOptions,
-    MoveDirection,
-    OutMode,
-} from "@tsparticles/engine";
-// import { loadAll } from "@tsparticles/all"; // if you are going to use `loadAll`, install the "@tsparticles/all" package too.
-// import { loadFull } from "tsparticles"; // if you are going to use `loadFull`, install the "tsparticles" package too.
-import { loadSlim } from "@tsparticles/slim"; // if you are going to use `loadSlim`, install the "@tsparticles/slim" package too.
-import { Box } from "@chakra-ui/react";
-// import { loadBasic } from "@tsparticles/basic"; // if you are going to use `loadBasic`, install the "@tsparticles/basic" package too.
+const Hero = () => {
+    return (
+        <Box display={"flex"} justifyContent={"center"} height={"calc(100vh - 180px)"} alignItems={"center"} >
+            <Container maxW={"container.xl"}>
+                <Heading color={"white"} textAlign={"center"}>
+                    Artificial Intelligence in Technology Lab
+                </Heading>
+                <List spacing={2} textAlign={"center"} color={"white"} mt={4} fontSize={"xl"}>
+                    <ListItem>
+                        <ListIcon as={HiMiniArrowRightCircle} color={'white'} />
+                        Casual Artificial Intelligence
+                    </ListItem>
+                    <ListItem>
+                        <ListIcon as={HiMiniArrowRightCircle} color={'white'} />
+                        Transportation Safety and Operations
+                    </ListItem>
+                    <ListItem>
+                        <ListIcon as={HiMiniArrowRightCircle} color={'white'} />
+                        Infrastructure readiness for disruptive technologies
+                    </ListItem>
+                </List>
+            </Container>
+        </Box >
+    )
+}
 
-export const Hero = () => {
-    const [init, setInit] = useState(false);
-
-    // this should be run only once per application lifetime
-    useEffect(() => {
-        initParticlesEngine(async (engine) => {
-            // you can initiate the tsParticles instance (engine) here, adding custom shapes or presets
-            // this loads the tsparticles package bundle, it's the easiest method for getting everything ready
-            // starting from v2 you can add only the features you need reducing the bundle size
-            //await loadAll(engine);
-            //await loadFull(engine);
-            await loadSlim(engine);
-            //await loadBasic(engine);
-        }).then(() => {
-            setInit(true);
-        });
-    }, []);
-
-    const particlesLoaded = async (container?: Container): Promise<void> => {
-        console.log(container);
-    };
-
-    const options: ISourceOptions = useMemo(
-        () => ({
-            background: {
-                color: {
-                    value: "#2b6cb0",
-                },
-            },
-            fpsLimit: 1120,
-            interactivity: {
-                events: {
-                    onClick: {
-                        enable: true,
-                        mode: "push",
-                    },
-                    onHover: {
-                        enable: true,
-                        mode: "repulse",
-                    },
-                },
-                modes: {
-                    push: {
-                        quantity: 4,
-                    },
-                    repulse: {
-                        distance: 100,
-                        duration: 1,
-                    },
-                },
-            },
-            particles: {
-                color: {
-                    value: "#ffffff",
-                },
-                links: {
-                    color: "#ffffff",
-                    distance: 150,
-                    enable: true,
-                    opacity: 0.5,
-                    width: 1,
-                },
-                move: {
-                    direction: MoveDirection.none,
-                    enable: true,
-                    outModes: {
-                        default: OutMode.out,
-                    },
-                    random: false,
-                    speed: 2,
-                    straight: false,
-                },
-                number: {
-                    density: {
-                        enable: true,
-                    },
-                    value: 80,
-                },
-                opacity: {
-                    value: 0.5,
-                },
-                shape: {
-                    type: "circle",
-                },
-                size: {
-                    value: { min: 1, max: 5 },
-                },
-            },
-            detectRetina: true,
-        }),
-        [],
-    );
-
-    if (init) {
-        return (
-            // Wrapping Particles in a Box with zIndex to control stacking
-            <Box position="absolute" zIndex={-1}>
-                <Particles
-                    id="tsparticles"
-                    particlesLoaded={particlesLoaded}
-                    options={options}
-                    style={{ position: "absolute", top: 0, left: 0, zIndex: -1 }}
-                />
-            </Box>
-
-        );
-    }
-
-    return <></>;
-};
+export default Hero;
